qesteidutil: fix compiler warnings
git-svn-id: https://svn.eesti.ee/projektid/idkaart/qesteidutil/trunk@3787 c45a0044-e149-41c6-9d26-0409f48354e9

+++ src/DiagnosticsDialog.cpp
@@ -63,7 +63,7 @@ QString DiagnosticsDialog::getReaderInfo()
 					ByteVec certBytes = card.getAuthCert();
 					if( certBytes.size() )
 					{
-						QSslCertificate cert = QSslCertificate( QByteArray((char *)&certBytes[0], certBytes.size()), QSsl::Der );
+						QSslCertificate cert = QSslCertificate( QByteArray((char *)&certBytes[0], (int)certBytes.size()), QSsl::Der );
 						if ( cert.isValid() && SslCertificate( cert ).isDigiID() )
 							id = cert.subjectInfo( "serialNumber" );
 					}
